pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "stratagem"
version = "0.1.0"
description = "Market research agent with document processing, web scraping, and financial analysis"
requires-python = ">=3.12"
license = "MIT"

dependencies = [
    "claude-agent-sdk>=0.1.48",
    "httpx>=0.28.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pypdf>=4.3.0",
    "pdfplumber>=0.11.0",
    "openpyxl>=3.1.4",
    "python-pptx>=1.0.0",
    "edgartools>=3.0.0",
    "Pillow>=11.0.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[project.scripts]
stratagem = "stratagem.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/stratagem"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["network: requires network access"]
